# swagger_local.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ecommerce Api server for admin and web",
    "version": "1.0.0",
    "title": "Ecommerce"
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/admin/auth/login/admin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login to admin system",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Admin username and password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/member/admin": {
      "get": {
        "tags": [
          "admin member"
        ],
        "summary": "Get all admins",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/member/admin/{adminId}": {
      "get": {
        "tags": [
          "admin member"
        ],
        "summary": "Find admin by id",
        "description": "Returns a single admin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "put": {
        "tags": [
          "admin member"
        ],
        "summary": "update admin by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "member update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "delete": {
        "tags": [
          "admin member"
        ],
        "summary": "Deletes an admin",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/admin/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get all products",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "post a product",
        "description": "",
        "operationId": "",
        "consumes": [
          "multipart/form-date"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "name",
            "description": "product name",
            "required": true
          },
          {
            "in": "formData",
            "name": "cartDescription",
            "description": "cart",
            "required": true
          },
          {
            "in": "formData",
            "name": "longDescription",
            "description": "long Desc",
            "required": true
          },
          {
            "in": "formData",
            "name": "shortDescription",
            "description": "short desc",
            "required": true
          },
          {
            "in": "formData",
            "name": "location",
            "description": "product name",
            "required": true
          },
          {
            "in": "formData",
            "name": "stock",
            "description": "product name",
            "required": true
          },
          {
            "in": "formData",
            "name": "price",
            "description": "price",
            "required": true
          },
          {
            "in": "formData",
            "name": "visible",
            "description": "visibility",
            "required": false
          },
          {
            "in": "formData",
            "name": "categoryID",
            "description": "category id",
            "required": true
          },
          {
            "in": "formData",
            "name": "sellerID",
            "description": "seller id",
            "required": true
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file",
            "description": "thumbnail image",
            "required": true
          },
          {
            "in": "formData",
            "name": "productImage",
            "type": "file",
            "description": "main image",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get this product",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "put": {
        "tags": [
          "product"
        ],
        "summary": "update product info",
        "description": "",
        "operationId": "",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of method to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "name",
            "description": "product name",
            "required": true
          },
          {
            "in": "formData",
            "name": "cartDescription",
            "description": "cart",
            "required": true
          },
          {
            "in": "formData",
            "name": "longDescription",
            "description": "long Desc",
            "required": true
          },
          {
            "in": "formData",
            "name": "shortDescription",
            "description": "short desc",
            "required": true
          },
          {
            "in": "formData",
            "name": "location",
            "description": "product name",
            "required": true
          },
          {
            "in": "formData",
            "name": "stock",
            "description": "product name",
            "required": true
          },
          {
            "in": "formData",
            "name": "price",
            "description": "price",
            "required": true
          },
          {
            "in": "formData",
            "name": "visible",
            "description": "visibility",
            "required": false
          },
          {
            "in": "formData",
            "name": "categoryID",
            "description": "category id",
            "required": true
          },
          {
            "in": "formData",
            "name": "sellerID",
            "description": "seller id",
            "required": true
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file",
            "description": "thumbnail image",
            "required": false
          },
          {
            "in": "formData",
            "name": "productImage",
            "type": "file",
            "description": "main image",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "delete product",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/product/image/{imageId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get thsimage",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "description": "image id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/product/thumbnail/{imageId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get thsimage",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "description": "image id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/product/image/{imageId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get thsimage",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "description": "image id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/product/thumbnail/{imageId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get thsimage",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "imageId",
            "in": "path",
            "description": "image id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/product/category": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get all category",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "post category",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "add category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/product/category/{categoryId}": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "update category",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of category to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "add category",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "delete product category",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID of product to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get all products",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get products",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "categoryName": {
          "type": "string",
          "example": "Game Points"
        }
      }
    },
...
    "RegisterAdmin": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himalhs1@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "himal123"
        },
        "email": {
          "type": "string",
          "example": "himal123"
        },
        "roles": {
          "type": "string",
          "example": "himal123"
        },
        "fname": {
          "type": "string",
          "example": "himal123"
        },
        "lname": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "RegisterAdmin"
      }
    },
...
    "Auth": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himal"
        },
        "password": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "Pet"
      }
    }
  },
...
}